Added FAT12/16 and FAT32 support.

+++ exfatClass.cpp
@@ -47,7 +47,7 @@ DWORD exfatClass::getTotalClusters(){
 	return totalClusters;
 }
 
-DWORD exfatClass::getFsClustersOffset(){
+double exfatClass::getFsClustersOffset(){
 	return firstDataSector / sectorsPerCluster;
 }
